django_telegrambot_converter/requirements.txt
amqp==5.3.1
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Deprecated==1.2.18
Django==5.1.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-simple-history==3.7.0
django-timezone-field==7.1
Flask==3.0.3
idna==3.10
importlib-metadata==4.11.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
legacy==0.1.7
MarkupSafe==3.0.2
prompt_toolkit==3.0.50
pyTelegramBotAPI==4.23.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==4.0.0
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.1
telebot==0.0.5
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
